Move gdn_attention to tpu_inference/layers/common, remove unnecessary ops
