+++ tpu_inference/layers/vllm/ops/__init__.py
@@ -12,6 +12,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from tpu_inference.layers.vllm.ops import gdn_attention as gdn_attention
 from tpu_inference.layers.vllm.ops import \
     scaled_dot_product_attention as scaled_dot_product_attention
