+++ tpu_inference/models/vllm/vllm_model_wrapper.py
@@ -148,8 +148,6 @@ class VllmModelWrapper:
             needs_env=False,
         )
 
-        patch_ops.gdn_attention.apply_gated_delta_net_torch_ops_patch()
-
     def _apply_pp_patch(self):
         # patch `get_pp_group` in vLLM to jax's get_pp_group.
         import sys
